requirements_fixed.txt
# Core ML Libraries (tested with Python 3.11+)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Data Collection
requests>=2.31.0
netCDF4>=1.6.0
h5py>=3.8.0

# NASA/Earth Data APIs (updated version)
earthdata>=0.2.0
pydap>=3.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# API Backend
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
joblib>=1.3.0

# Geospatial
geopy>=2.4.0

# Model evaluation
imbalanced-learn>=0.11.0

# Note: Removed statsmodels and scikit-optimize temporarily due to Python 3.13 compatibility
